Fix: avoid creating empty FlatTensors (#115)
When enabling efficient memory sharing (EMS) technique, a Tensor to be offloaded might be very large to exceed the bucket_size_mb_in_memory_manager. In this case, the code in chatlearn/utils/flat_tensors.py#L255-265 will create a FlatTensors object with an empty bucket. This is not the expected behavior. This PR commit resolves the issue.